creating displaySocialLoginList method
